add data with private user login

+++ KOSIGNSTORE _iOS/Application/Main/Home/View/Home/NewUIDownloadTopic/Cell/AppDetailCell.swift
@@ -17,7 +17,8 @@ struct AppDetailCell: View {
     @State private  var showShareSheet      = false
     @State var listAppVersion      : AppDetail? = nil
     @State var items : [Any] = []
-    
+    @AppStorage("signIn") var isSignIn = false
+    @State private  var showAlert      = false
     // MARK: - Custom PopSheet
     let sheetConfiguration: SheetConfiguration = SheetConfiguration (
         
@@ -73,42 +74,69 @@ struct AppDetailCell: View {
                     }
                     // MARK: - Button Download --> FILE.pdf
                     VStack {
-                        Button(action: {
+                        //check with sign up with google
+                        if !isSignIn {
+                            Button(action: {
+                                self.showAlert = true
 
-                            if let localPath = listAppVersion?.pathFile, !localPath.isEmpty {
-                                let fileName = (localPath as NSString).lastPathComponent
-                                let destinationURL = FileManager.default.temporaryDirectory.appendingPathComponent(fileName)
-                                
-                                // Check if the file already exists
-                                if FileManager.default.fileExists(atPath: destinationURL.path) {
+                            }, label: {
+                                Text("Download")
+                                    .font(.customFont(font: .Rubik, style: .bold, size: .h4))
+                                    .foregroundColor(Color.white)
+                                    .frame(height: 50)
+                                    .frame(maxWidth: .infinity)
+                                    .background(Color.gray)
+                                    .cornerRadius(8.0)
+                            })
+                            .buttonStyle(PressableButtonStyle())
+                         
+
+                            .alert(isPresented: $showAlert, content: { () -> Alert in
+                                Alert(title: Text("Error").foregroundColor(Color("NewColor") ), message: Text("Please Login"), dismissButton: .default (
+                                    Text("Okay")
+                                        .foregroundColor(Color("NewColor"))
+                                ))
+                            })
+                            
+                        }else {
+                            Button(action: {
+
+                                if let localPath = listAppVersion?.pathFile, !localPath.isEmpty {
+                                    let fileName = (localPath as NSString).lastPathComponent
+                                    let destinationURL = FileManager.default.temporaryDirectory.appendingPathComponent(fileName)
+                                    
+                                    // Check if the file already exists
+                                    if FileManager.default.fileExists(atPath: destinationURL.path) {
+                                        do {
+                                            // If file exists, remove it first
+                                            try FileManager.default.removeItem(at: destinationURL)
+                                        } catch {
+                                            
+                                            print("Removing existing file failed with error: \(error)")
+                                        }
+                                    }
                                     do {
-                                        // If file exists, remove it first
-                                        try FileManager.default.removeItem(at: destinationURL)
+                                        let fileURL = URL(fileURLWithPath: localPath)
+                                        try FileManager.default.copyItem(at: fileURL, to: destinationURL)
+                                        items.append(destinationURL)
+                                        showShareSheet.toggle()
                                     } catch {
-                                        
-                                        print("Removing existing file failed with error: \(error)")
+                                        print("Copying local file failed with error: \(error)")
                                     }
                                 }
-                                do {
-                                    let fileURL = URL(fileURLWithPath: localPath)
-                                    try FileManager.default.copyItem(at: fileURL, to: destinationURL)
-                                    items.append(destinationURL)
-                                    showShareSheet.toggle()
-                                } catch {
-                                    print("Copying local file failed with error: \(error)")
-                                }
-                            }
-                        }, label: {
-                            Text("Download")
-                                .font(.customFont(font: .Rubik, style: .bold, size: .h4))
-                                .foregroundColor(Color.white)
-                                .frame(height: 50)
-                                .frame(maxWidth: .infinity)
-                                .background(Color("ColorDownload"))
-                                .cornerRadius(8.0)
-                        })
-                        .buttonStyle(PressableButtonStyle())
-                        
+                            }, label: {
+                                Text("Download")
+                                    .font(.customFont(font: .Rubik, style: .bold, size: .h4))
+                                    .foregroundColor(Color.white)
+                                    .frame(height: 50)
+                                    .frame(maxWidth: .infinity)
+                                    .background(Color("ColorDownload"))
+                                    .cornerRadius(8.0)
+                            })
+                            .buttonStyle(PressableButtonStyle())
+                            
+
+                        }
                     }
                 }
             }
